style: full width row link
