+++ packages/frontend/src/components/common/ResourceList/ResourceList.styles.ts
@@ -1,5 +1,4 @@
 import { Card, Colors, H3, Tag } from '@blueprintjs/core';
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 const paddingX = 20;
@@ -44,24 +43,4 @@ export const ResourceTag = styled(Tag)`
     font-weight: bold;
     background-color: ${Colors.LIGHT_GRAY2};
     color: ${Colors.DARK_GRAY1};
-`;
-
-export const ResourceLink = styled(Link)`
-    font-size: 13px;
-    font-weight: 600;
-    color: ${Colors.DARK_GRAY4};
-
-    &:hover {
-        color: ${Colors.DARK_GRAY1};
-    }
-`;
-
-export const ResourceSpaceLink = styled(Link)`
-    font-size: 13px;
-    font-weight: 500;
-    color: ${Colors.GRAY2};
-
-    &:hover {
-        color: ${Colors.GRAY1};
-    }
 `;
