+++ packages/frontend/src/components/common/ResourceList/ResourceTable/ResourceTable.styles.ts
@@ -1,4 +1,5 @@
 import { Colors, HTMLTable } from '@blueprintjs/core';
+import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 const paddingX = 20;
@@ -11,15 +12,22 @@ export const StyledTHead = styled.thead`
     background-color: ${Colors.LIGHT_GRAY5};
 `;
 
-export const StyledTr = styled.tr``;
+export const StyledTr = styled.tr`
+    position: relative;
+
+    :hover {
+        background-color: ${Colors.LIGHT_GRAY5}75;
+    }
+
+    > td > *:not(a.full-row-link) {
+        position: relative;
+        z-index: 2;
+    }
+`;
 
 export const StyledTBody = styled.tbody`
     ${StyledTr}:not(:last-child) {
         box-shadow: inset 0 -1px 0 0 ${Colors.LIGHT_GRAY3} !important;
-    }
-
-    ${StyledTr}:hover {
-        background-color: ${Colors.LIGHT_GRAY5};
     }
 `;
 
@@ -67,4 +75,37 @@ export const Flex = styled.div`
 
 export const Spacer = styled.div<SpacerProps>`
     width: ${(props) => props.$width}px;
+`;
+
+export const NoLinkContainer = styled.div`
+    pointer-events: none;
+`;
+
+export const ResourceLink = styled(Link)`
+    font-size: 13px;
+    font-weight: 600;
+    color: ${Colors.DARK_GRAY4};
+
+    &:hover {
+        color: ${Colors.DARK_GRAY1};
+    }
+
+    &.full-row-link {
+        position: absolute;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+        z-index: 1;
+    }
+`;
+
+export const ResourceSpaceLink = styled(Link)`
+    font-size: 13px;
+    font-weight: 500;
+    color: ${Colors.GRAY2};
+
+    &:hover {
+        color: ${Colors.GRAY1};
+    }
 `;
